refactor(ScoreAgent): revese course on mocking
the original plan was to use the new public getters to be mocked setup ScoreAgent
with the correct private variable values. but upon consideration a black
box approach would be less dependent on any implementation, so calculate
winning would not use any getter anymore, instead it will access the
variables directly.